TimeEntries no longer require hours to be an integer. Hours must be greater than 0 and at most 24.

+++ db/migrate/20090410004750_create_time_entries.rb
@@ -4,7 +4,7 @@ class CreateTimeEntries < ActiveRecord::Migration
       t.references :timesheet
       t.date :entry_date
       t.references :activity_type
-      t.integer :hours
+      t.float :hours
       t.string :description
 
       t.timestamps
